feat(utxo-lib): move outputScripts to bitgo subpackage
Issue: BG-7478
